feat: 조회수 Long 타입 && 단일 조회 시 viewCount +1, 1분 이후, db 값을 받아오는 로직 추가

+++ src/main/java/com/economy/community/controller/PostController.java
@@ -108,15 +108,15 @@ public class PostController {
 
     // 게시글 조회수 증가
     @GetMapping("/{id}/view")
-    public ResponseEntity<Integer> postView(@PathVariable Long id) {
-        int updatedViewCount = postViewCountService.incrementPostViewCount(id);
+    public ResponseEntity<Long> postView(@PathVariable Long id) {
+        Long updatedViewCount = postViewCountService.incrementPostViewCount(id);
         return ResponseEntity.ok(updatedViewCount);
     }
 
     // 게시글 조회수 가져오기
     @GetMapping("/{id}/viewCount")
-    public ResponseEntity<Integer> getPostViewCount(@PathVariable Long id) {
-        int viewCount = postViewCountService.getPostViewCount(id);
+    public ResponseEntity<Long> getPostViewCount(@PathVariable Long id) {
+        Long viewCount = postViewCountService.getPostViewCount(id);
         return ResponseEntity.ok(viewCount);
     }
 }
